tests/wpt/web-platform-tests/css/css-position/parsing/inset-before-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module Level 3: parsing inset-before with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#propdef-inset-before">
<meta name="assert" content="inset-before supports the full grammar 'auto | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("inset-before", "auto");

test_valid_value("inset-before", "-10px");
test_valid_value("inset-before", "-20%");
test_valid_value("inset-before", "calc(2em + 3ex)");
</script>
</body>
</html>
